perf(frontend): Reduce Props interface

+++ packages/backend/scripts/copy-files.ts
@@ -4,7 +4,9 @@ import * as fs from 'fs';
 
 export const copyFiles = ({ pluginsPath }: { pluginsPath: string }) => {
   const currentPathToCoreAdmin = path.join(pluginsPath, 'core', 'admin');
-  fs.rmSync(currentPathToCoreAdmin, { recursive: true });
+  if (fs.existsSync(currentPathToCoreAdmin)) {
+    fs.rmSync(currentPathToCoreAdmin, { recursive: true });
+  }
   fs.mkdirSync(currentPathToCoreAdmin, { recursive: true });
 
   // Copy core plugin
